+++ snooplyze/page/page.py
@@ -1,6 +1,8 @@
 from dataclasses import dataclass
+from typing import Optional
 
 @dataclass
 class Page:
     name : str
     url : str
+    description: Optional[str]