+++ snooplyze/parser/__init__.py
@@ -1,10 +1,9 @@
-from .base_parser import ParserBase
-from .parser_registry import PARSER_REGISTRY
+from .base_parser import ParserBase, ParserType
 from .generic_parsers import AllDocumentParser, MainElementParser, DivClassParser
 from .custom_parsers import DuckDbBlogParser, MicrosoftAzureVirtualMachinesBlogParser, MicrosoftAzureAzureSQLDatabaseBlogParser
 
 # Now, from persistence import PARSER_REGISTRY will work
-__all__ = ["ParserBase", "PARSER_REGISTRY",
+__all__ = ["ParserBase", "ParserType",
            
            "AllDocumentParser", "MainElementParser", "DivClassParser",