+++ snooplyze/parser/base_parser.py
@@ -1,17 +1,14 @@
 from abc import ABC, abstractmethod
+from enum import Enum
+
+class ParserType(str, Enum):
+    ALL_DOCUMENT = "alldocument"
+    MAIN_ELEMENT = "mainelement"
+    DIV_CLASS = "divclass"
+
 
 class ParserBase(ABC):
     
     @abstractmethod
     def parse(self, text_to_parse):
         pass
-
-    @classmethod
-    @abstractmethod
-    def validate(cls, data: dict) -> None:
-        pass
-
-    @classmethod
-    def from_dict(cls, data: dict):
-        cls.validate(data)
-        return cls(**data)